Help function getAutoGroupCommandStr moved in new file src/Mod/Part/Gui/Utils.cpp
