+++ src/Mod/Part/Gui/DlgPrimitives.cpp
@@ -53,6 +53,7 @@
 #include <Mod/Part/App/FeaturePartBox.h>
 #include <Mod/Part/App/FeaturePartCircle.h>
 #include <Mod/Part/App/Tools.h>
+#include "Utils.h"
 
 #include "DlgPrimitives.h"
 #include "ui_DlgPrimitives.h"
@@ -64,23 +65,6 @@ using namespace PartGui;
 namespace PartGui
 {
 
-QString getAutoGroupCommandStr(QString objectName)
-// Helper function to get the python code to add the newly created object to the active Part object
-// if present
-{
-    App::Part* activePart = Gui::Application::Instance->activeView()->getActiveObject<App::Part*>(
-        "part"
-    );
-    if (activePart) {
-        QString activeObjectName = QString::fromUtf8(activePart->getNameInDocument());
-        return QStringLiteral(
-                   "App.ActiveDocument.getObject('%1\')."
-                   "addObject(App.ActiveDocument.getObject('%2\'))\n"
-        )
-            .arg(activeObjectName, objectName);
-    }
-    return QStringLiteral("# Object %1 created at document root").arg(objectName);
-}
 
 const char* gce_ErrorStatusText(gce_ErrorType et)
 {
@@ -2308,7 +2292,7 @@ void DlgPrimitives::tryCreatePrimitive(const QString& placement)
     QString prim = tr("Create %1").arg(ui->PrimitiveTypeCB->currentText());
     Gui::Application::Instance->activeDocument()->openCommand(prim.toUtf8());
     Gui::Command::runCommand(Gui::Command::Doc, cmd.toUtf8());
-    Gui::Command::runCommand(Gui::Command::Doc, getAutoGroupCommandStr(name).toUtf8());
+    Gui::Command::runCommand(Gui::Command::Doc, PartGui::getAutoGroupCommandStr(name).toUtf8());
     Gui::Application::Instance->activeDocument()->commitCommand();
     Gui::Command::runCommand(Gui::Command::Doc, "App.ActiveDocument.recompute()");
     Gui::Command::runCommand(Gui::Command::Gui, "Gui.SendMsgToActiveView(\"ViewFit\")");
